+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version XXX:
 
+* Update WebSocket examples to set TLS SNI.
 * Add handler tracking locations to websocket.
 * Add handler tracking locations to tcp teardown.
 * Eliminate spurious uninitialised variable warning in detect_ssl.